requirements.txt
certifi==2019.3.9
Django==2.1.5
django-crispy-forms==1.7.2
django-taggit==0.24.0
django-webp-converter==0.1.2
gunicorn==19.9.0
httpagentparser==1.8.2
mock==2.0.0
pbr==5.1.3
Pillow==6.0.0
psycopg2==2.7.7
pytz==2018.9
six==1.12.0
urllib3==1.24.1
